refactor(vm): moved `ExtendedVariant` to `Helper` folder
