+++ VM/include/RigCVM/Helper/ExtendedVariant.hpp
